feat(nvim): disable kind auto_brackets for LaTeX
In *tex kind auto_brackets causes insertion of brackets and that's
because of the way texlab language server is setup. It perceives
commands as functions. For now it seems there's no known particular
mechanism for it to distinguish between commands that accept arguments
and those that don't accept.

Here's the link to the discussions over on the GitHub:
- https://github.com/Saghen/blink.cmp/discussions/904#discussioncomment-11763963
- https://github.com/latex-lsp/texlab/discussions/1307