doc(git/.gitconfig): explain 'autosquash = true'
Now that I know what --autosquash does, I added some explanation
comments to the global config where I set the autosquash = true under
the rebase.
--autosquash simply means that during rebase git rearranges the commits
that start with 'fixup!' or 'squash!' to the right place in the todo
list, making it a bit more convenient to work with interactive rebase.